requirements.txt
numpy
opencv-contrib-python
requests
azure-cognitiveservices-vision-computervision
Pillow
socketIO-client-nexus==0.7.2